+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New Features
 - PR #235: NearestNeighbors MG Support
+- PR #290: NearestNeighbors numpy MG Support
 
 
 ## Improvements
@@ -29,6 +30,7 @@
 - PR #253: Fix for issue #250-reduce_rows_by_key failed memcheck for small nkeys
 - PR #269: LinearRegression, Ridge Python docs update and cleaning
 - PR #275: Kmeans use of faster gpu_matrix
+- PR #288: Add n_neighbors to NearestNeighbors constructor
 
 ## Bug Fixes